+++ CHANGELOG.md
@@ -1,7 +1,8 @@
-## Unreleased
+## Unreleased (future 0.6.0)
 
   * Improve example server to make it more useful for local testing & development. [#17](https://github.com/alexdean/as2/pull/17)
-  * Server can examine HTTP headers to determine desired MIC algorithm and respond accordingly.
+  * Support `Content-Tranfer-Encoding: binary`. [#11](https://github.com/alexdean/as2/pull/11)
+  * Server can choose MIC algorithm based on HTTP `Disposition-Notification-Options` header. [#20](https://github.com/alexdean/as2/pull/20)
 
 ## 0.5.1, August 10, 2022